+++ src/app/src/components/SurveyDialogNotification.jsx
@@ -10,7 +10,7 @@ import attempt from 'lodash/attempt';
 import isError from 'lodash/isError';
 import moment from 'moment';
 
-const surveyURL = 'http://bit.ly/OARSurvey';
+const surveyURL = 'https://openapparelregistry.typeform.com/to/ue3mV6OD';
 
 const surveyDialogStyles = Object.freeze({
     containerStyles: Object.freeze({
@@ -25,31 +25,55 @@ const surveyDialogStyles = Object.freeze({
     }),
 });
 
-const InvisibleDiv = constant(<div style={{ display: 'none ' }} />);
-
-const surveyEndDate = moment('2019-11-10');
-const currentDate = moment();
+// We previously stored SURVEY_DIALOG_HAS_BEEN_DISPLAYED to determine
+// whether to show the survey popup. We are removing this key to avoid
+// confusion.
 const SURVEY_DIALOG_HAS_BEEN_DISPLAYED = 'SURVEY_DIALOG_HAS_BEEN_DISPLAYED';
-
-const trySetDialogHasBeenDisplayedToLocalStorage = () => attempt(
-    () => window.localStorage.setItem(
-        SURVEY_DIALOG_HAS_BEEN_DISPLAYED,
-        SURVEY_DIALOG_HAS_BEEN_DISPLAYED,
-    ),
-);
-
-const dialogEntryIsNotSavedInLocalStorage = () => {
-    const result = attempt(
+const removeSurveyDialogueHasBeenDisplayed = () => {
+    const surveyDialogueHasBeenDisplayed = attempt(
         () => window.localStorage.getItem(SURVEY_DIALOG_HAS_BEEN_DISPLAYED),
     );
 
-    return isError(result) ? true : !result;
+    if (!isError(surveyDialogueHasBeenDisplayed) &&
+        !!surveyDialogueHasBeenDisplayed) {
+        attempt(
+            () => window.localStorage.removeItem(
+                SURVEY_DIALOG_HAS_BEEN_DISPLAYED,
+            ),
+        );
+    }
+};
+removeSurveyDialogueHasBeenDisplayed();
+
+const InvisibleDiv = constant(<div style={{ display: 'none ' }} />);
+
+const dateFormat = 'YYYY-MM-DD';
+const surveyStartDate = moment('2020-11-5', dateFormat);
+const surveyEndDate = moment('2020-12-13', dateFormat);
+const currentDate = moment();
+
+const SURVEY_DIALOG_DISMISSED_DATE = 'SURVEY_DIALOG_DISMISSED_DATE';
+
+const trySetDialogHasBeenDisplayedToLocalStorage = () => attempt(
+    () => window.localStorage.setItem(
+        SURVEY_DIALOG_DISMISSED_DATE,
+        currentDate.format(dateFormat),
+    ),
+);
+
+const dialogDateIsNotSavedAndCurrentInLocalStorage = () => {
+    const result = attempt(
+        () => window.localStorage.getItem(SURVEY_DIALOG_DISMISSED_DATE),
+    );
+
+    return isError(result) || moment(result).isBefore(surveyStartDate) || !result;
 };
 
 export default function SurveyDialogNotification() {
     const [dialogIsOpen, setDialogIsOpen] = useState(
+        currentDate.isSameOrAfter(surveyStartDate) &&
         currentDate.isBefore(surveyEndDate) &&
-            dialogEntryIsNotSavedInLocalStorage(),
+            dialogDateIsNotSavedAndCurrentInLocalStorage(),
     );
 
     const closeDialog = () => {