+++ src/django/api/models.py
@@ -1764,8 +1764,9 @@ class ApiLimit(models.Model):
     history = HistoricalRecords()
 
     def __str__(self):
-        return ('ApiLimit {id} - Contributor {contributor_id} '
-                'limit {monthly_limit}').format(**self.__dict__)
+        return 'ApiLimit {} - {} ({}) - limit {}'.format(
+            self.id, self.contributor.name, self.contributor.id,
+            self.monthly_limit)
 
 
 class ApiBlock(models.Model):
@@ -1818,3 +1819,37 @@ class ApiBlock(models.Model):
     def __str__(self):
         return ('ApiBlock {id} - Contributor {contributor_id} until '
                 '{until}').format(**self.__dict__)
+
+
+class ContributorNotifications(models.Model):
+    """
+    Records notifications sent to contributors.
+    """
+    contributor = models.OneToOneField(
+        'Contributor',
+        null=False,
+        on_delete=models.CASCADE,
+        help_text='The contributor to whom the notification was sent.'
+    )
+    api_limit_warning_sent_on = models.DateTimeField(
+        null=True,
+        help_text='When a limit warning was sent to the contributor.')
+    api_limit_exceeded_sent_on = models.DateTimeField(
+        null=True,
+        help_text='When a limit exceeded notice was sent to the contributor.')
+    api_grace_limit_exceeded_sent_on = models.DateTimeField(
+        null=True,
+        help_text=('When a grace limit exceeded notice was sent '
+                   'to the contributor.'))
+
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    history = HistoricalRecords()
+
+    def __str__(self):
+        return ('ContributorNotification {id} - Contributor {contributor_id} '
+                'Warning: {api_limit_warning_sent_on}, '
+                'Exceeded: {api_limit_exceeded_sent_on}, '
+                'Grace exceeded: {api_grace_limit_exceeded_sent_on} '
+                ).format(**self.__dict__)