+++ src/django/oar/settings.py
@@ -172,6 +172,7 @@ MIDDLEWARE = [
     'simple_history.middleware.HistoryRequestMiddleware',
     'waffle.middleware.WaffleMiddleware',
     'api.middleware.RequestLogMiddleware',
+    'api.middleware.RequestMeterMiddleware',
 ]
 
 ROOT_URLCONF = 'oar.urls'
@@ -233,6 +234,9 @@ AUTH_PASSWORD_VALIDATORS = [
 
 AUTH_USER_MODEL = 'api.User'
 
+# Api Limits
+API_FREE_REQUEST_LIMIT = 50
+
 # Logging
 # https://docs.djangoproject.com/en/2.1/topics/logging/
 
@@ -277,6 +281,9 @@ else:
 
 DEFAULT_FROM_EMAIL = os.getenv(
     'DEFAULT_FROM_EMAIL', 'noreply@staging.openapparel.org')
+
+NOTIFICATION_EMAIL_TO = os.getenv(
+    'NOTIFICATION_EMAIL_TO', 'notification@example.com')
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.0/howto/static-files/